+++ myalbum-service-provider-user/src/main/java/cn.yan_wm.myalbum.service.provider.user/service/UserFanService.java
@@ -1,8 +1,10 @@
 package cn.yan_wm.myalbum.service.provider.user.service;
 
+import cn.yan_wm.myalbum.commons.domain.TbImage;
 import cn.yan_wm.myalbum.commons.domainExtend.user.UserAttentionExtend;
 import cn.yan_wm.myalbum.commons.domainExtend.user.UserFanExtend;
 import cn.yan_wm.myalbum.commons.model.DataSet;
+import tk.mybatis.mapper.entity.Example;
 import tk.mybatis.page.Page;
 
 
@@ -20,7 +22,7 @@ public interface UserFanService{
      * @param fanId
      * @return
      */
-    int addFan(Long userId, Long fanId);
+    int addFan(Integer userId, Integer fanId);
 
     /**
      * 删除粉丝
@@ -28,7 +30,7 @@ public interface UserFanService{
      * @param fanId
      * @return
      */
-    int deleteFan(Long userId, Long fanId);
+    int deleteFan(Integer userId, Integer fanId);
 
     /**
      * 修改粉丝备注
@@ -37,7 +39,7 @@ public interface UserFanService{
      * @param note
      * @return
      */
-    int updateFanNote(Long userId, Long fanId, String note);
+    int updateFanNote(Integer userId, Integer fanId, String note);
 
     /**
      * 查询粉丝信息
@@ -45,14 +47,14 @@ public interface UserFanService{
      * @param fanId
      * @return
      */
-    UserFanExtend getByFanId(Long userId, Long fanId);
+    UserFanExtend getByFanId(Integer userId, Integer fanId);
 
     /**
      * 通过用户id 查询所有的粉丝
      * @param userId
      * @return
      */
-    List<UserFanExtend> getByUserId(Long userId);
+    List<UserFanExtend> getByUserId(Integer userId);
 
     /**
      * 分页查询粉丝信息
@@ -60,5 +62,5 @@ public interface UserFanService{
      * @param page
      * @return
      */
-    DataSet<UserFanExtend> page(Long userId, Page page);
+    DataSet<UserFanExtend> page(Integer userId, Page page);
 }
